pyproject.toml
[project]
name = "hyperflask-users"
version = "0.3.1"
description = "Authentication and users management for Hyperflask"
authors = [
    {"name" = "Maxime Bouroumeau-Fuseau", email = "maxime.bouroumeau@gmail.com"}
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "hyperflask >=0.3.0",
    "flask-login >=0.6,<1.0",
    "passlib >=1.7,<2.0",
]

[project.urls]
Repository = "https://github.com/hyperflask/hyperflask-users"

[project.entry-points."hyperflask.extensions"]
users = "hyperflask_users:Users"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.4",
]

[tool.ruff]
line-length = 100

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"
